+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K5GVCVER21EE2TMA21173KQ1"
+        "01K5P8H8AKT5S0DDEFFVAY2YX1"
     )
     r"""The id of the resource"""