+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -59,5 +59,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-09-19T11:22:43.676Z")
+    created: Optional[datetime] = parse_datetime("2025-09-21T13:48:31.474Z")
     r"""The date and time the resource was created"""