+++ packages/orq-rc/src/orq_ai_sdk/models/getagentop.py
@@ -230,7 +230,7 @@ class GetAgentKnowledgeBases(BaseModel):
     configuration: GetAgentConfiguration
     r"""Defines the configuration settings which can either be for a user message or a text entry."""
 
-    id: Optional[str] = "01K5GVCVA2R75KYWJFYSTPN41F"
+    id: Optional[str] = "01K5P8H85R69VC35GMP021YCW9"
     r"""The id of the resource"""