+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -127,7 +127,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K5GVCVEQ9YM6TP83158SCKZM"
+        "01K5P8H8AH2NM9ZDY3P736VXJW"
     )
     r"""The id of the resource"""