+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -86,7 +86,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K5GVCVEQEQ059H623K94KN22"
+        "01K5P8H8AJZE076RPPRHP7W0X9"
     )
     r"""The id of the resource"""