+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -505,7 +505,7 @@ class Twelve(BaseModel):
     http: HTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K5GVCVAT62Q2KSZCYQ4ZMK81"
+        "01K5P8H86HXDY2EG6M545NB9XM"
     )
 
     status: Optional[ToolsStatus] = "live"